Fix package paths and add missing utils
