+++ lib/pages/pixel_control_page.dart
@@ -1,25 +1,30 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import '../ble_helper.dart';
-import '../Utils/style.dart'; // Buton stil fonksiyonları
+import '../utils/ble_helper.dart';
+import '../utils/style.dart'; // Buton stil fonksiyonları
 import '../utils/colors.dart'; // GetX tema renkleri
 
-class SliderPage extends StatefulWidget {
-  const SliderPage({Key? key}) : super(key: key);
+class SinglePixelControlPage extends StatefulWidget {
+  const SinglePixelControlPage({Key? key}) : super(key: key);
 
   @override
-  _SliderPageState createState() => _SliderPageState();
+  _SinglePixelControlPageState createState() => _SinglePixelControlPageState();
 }
 
-class _SliderPageState extends State<SliderPage> {
+class _SinglePixelControlPageState extends State<SinglePixelControlPage> {
   final appColorsController = Get.put(AppColorsController());
+
+  // Kullanıcının ayarlayacağı piksel numarası (1-indexli)
+  final TextEditingController _pixelController = TextEditingController(
+    text: "1",
+  );
 
   double red = 0;
   double green = 0;
   double blue = 0;
-  double white = 0; // White slider için değişken
-  double brightness = 255; // Ön izleme kutusundaki parlaklık değeri
+  double white = 0; // Sadece RGBW modunda kullanılacak.
+  double brightness = 255; // Parlaklık değeri
 
   bool isSending = false;
   Timer? _debounce;
@@ -34,9 +39,17 @@ class _SliderPageState extends State<SliderPage> {
     });
   }
 
-  /// BLE cihazına RGB veya RGBW komutunu gönderir.
+  /// BLE üzerinden sadece seçili pikselin renk değerlerini gönderir.
   Future<void> sendColorCommand() async {
-    if (!isSending) return;
+    int pixelNumber = int.tryParse(_pixelController.text) ?? 0;
+    if (pixelNumber <= 0) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text("Lütfen geçerli bir piksel numarası girin."),
+        ),
+      );
+      return;
+    }
 
     int r = red.toInt();
     int g = green.toInt();
@@ -44,69 +57,96 @@ class _SliderPageState extends State<SliderPage> {
     int p = brightness.toInt();
 
     String command;
-    if (BLEHelper.mode.value == "RGBW") {
+    // Mod kontrolünü boşlukları silip büyük harf olarak yapıyoruz.
+    bool isRGBW = (BLEHelper.mode.value.trim().toUpperCase() == "RGBW");
+    if (isRGBW) {
       int w = white.toInt();
-      command = "RGBW:$r,$g,$b,$w,$p";
+      command = "PIXCTRL:$pixelNumber:$r,$g,$b,$w,$p";
     } else {
-      command = "RGB:$r,$g,$b,$p";
+      command = "PIXCTRL:$pixelNumber:$r,$g,$b,$p";
     }
 
-    List<int> bytes = command.codeUnits;
+    print("Gönderilen piksel komutu: $command");
+
     if (BLEHelper.connectedCharacteristic != null) {
       try {
-        await BLEHelper.connectedCharacteristic!.write(bytes);
-        print("Gönderilen komut: $command");
+        await BLEHelper.connectedCharacteristic!.write(command.codeUnits);
       } catch (e) {
-        print("Renk komutu gönderme hatası: $e");
+        print("PIXCTRL komutu gönderme hatası: $e");
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text("Renk komutu gönderilemedi")),
+            const SnackBar(content: Text("PIXCTRL komutu gönderilemedi")),
           );
         }
-        // Hata alınırsa bağlı cihazı temizle.
-        BLEHelper.clearConnection();
       }
     } else {
-      print("Bağlı BLE cihazı yok!");
       if (mounted) {
         ScaffoldMessenger.of(
           context,
         ).showSnackBar(const SnackBar(content: Text("Bağlı BLE cihazı yok!")));
       }
-      BLEHelper.clearConnection();
     }
   }
 
   @override
+  void dispose() {
+    _debounce?.cancel();
+    _pixelController.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
+    // Mod kontrolü için büyük harfe çevirerek kontrol ediyoruz.
+    bool isRGBW = (BLEHelper.mode.value.trim().toUpperCase() == "RGBW");
     return Scaffold(
       backgroundColor: appColorsController.blackColor.value,
+      appBar: AppBar(
+        backgroundColor: appColorsController.blackColor.value,
+        // Geri tuşu eklenmiş durumda (AppBar'ın otomatik back buttonu çalışacaktır)
+        iconTheme: IconThemeData(color: appColorsController.whiteColor.value),
+        title: const Text(
+          "Tek Piksel Kontrol",
+          style: TextStyle(color: Colors.white),
+        ),
+      ),
       body: SafeArea(
         child: SingleChildScrollView(
           padding: const EdgeInsets.all(16.0),
           child: Column(
             children: [
-              const SizedBox(height: 50),
-              // Renk önizleme kutusu
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 24),
-                child: Container(
-                  width: double.infinity,
-                  height: 150,
-                  decoration: BoxDecoration(
-                    color: Color.fromARGB(
-                      brightness.toInt(),
-                      red.toInt(),
-                      green.toInt(),
-                      blue.toInt(),
-                    ),
-                    borderRadius: BorderRadius.circular(12),
-                    border: Border.all(color: Colors.white, width: 2),
+              const SizedBox(height: 20),
+              // Piksel numarası girilecek alan
+              TextField(
+                controller: _pixelController,
+                keyboardType: TextInputType.number,
+                style: TextStyle(color: appColorsController.whiteColor.value),
+                decoration: InputDecoration(
+                  labelText: "Piksel Numarası (1-indexli)",
+                  labelStyle: TextStyle(
+                    color: appColorsController.whiteColor.value,
                   ),
+                  border: const OutlineInputBorder(),
                 ),
               ),
-              const SizedBox(height: 40),
-              // Sliders bölümünü içeren Column
+              const SizedBox(height: 20),
+              // Renk önizleme kutusu
+              Container(
+                width: double.infinity,
+                height: 150,
+                decoration: BoxDecoration(
+                  color: Color.fromARGB(
+                    brightness.toInt(),
+                    red.toInt(),
+                    green.toInt(),
+                    blue.toInt(),
+                  ),
+                  borderRadius: BorderRadius.circular(12),
+                  border: Border.all(color: Colors.white, width: 2),
+                ),
+              ),
+              const SizedBox(height: 30),
+              // Renk ve parlaklık slider'ları
               Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
@@ -170,8 +210,8 @@ class _SliderPageState extends State<SliderPage> {
                       _onColorChanged();
                     },
                   ),
-                  // White slider, sadece mod "RGBW" ise gösterilir
-                  if (BLEHelper.mode.value == "RGBW") ...[
+                  // White slider, sadece RGBW modunda gösterilir
+                  if (isRGBW) ...[
                     Text(
                       "White: ${white.toInt()}",
                       style: const TextStyle(color: Colors.grey, fontSize: 18),
