+++ lib/pages/slider_page.dart
@@ -1,30 +1,25 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import '../ble_helper.dart';
-import '../Utils/style.dart'; // Buton stil fonksiyonları
+import '../utils/ble_helper.dart';
+import '../utils/style.dart'; // Buton stil fonksiyonları
 import '../utils/colors.dart'; // GetX tema renkleri
 
-class SinglePixelControlPage extends StatefulWidget {
-  const SinglePixelControlPage({Key? key}) : super(key: key);
+class SliderPage extends StatefulWidget {
+  const SliderPage({Key? key}) : super(key: key);
 
   @override
-  _SinglePixelControlPageState createState() => _SinglePixelControlPageState();
+  _SliderPageState createState() => _SliderPageState();
 }
 
-class _SinglePixelControlPageState extends State<SinglePixelControlPage> {
+class _SliderPageState extends State<SliderPage> {
   final appColorsController = Get.put(AppColorsController());
-
-  // Kullanıcının ayarlayacağı piksel numarası (1-indexli)
-  final TextEditingController _pixelController = TextEditingController(
-    text: "1",
-  );
 
   double red = 0;
   double green = 0;
   double blue = 0;
-  double white = 0; // Sadece RGBW modunda kullanılacak.
-  double brightness = 255; // Parlaklık değeri
+  double white = 0; // White slider için değişken
+  double brightness = 255; // Ön izleme kutusundaki parlaklık değeri
 
   bool isSending = false;
   Timer? _debounce;
@@ -39,17 +34,9 @@ class _SinglePixelControlPageState extends State<SinglePixelControlPage> {
     });
   }
 
-  /// BLE üzerinden sadece seçili pikselin renk değerlerini gönderir.
+  /// BLE cihazına RGB veya RGBW komutunu gönderir.
   Future<void> sendColorCommand() async {
-    int pixelNumber = int.tryParse(_pixelController.text) ?? 0;
-    if (pixelNumber <= 0) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text("Lütfen geçerli bir piksel numarası girin."),
-        ),
-      );
-      return;
-    }
+    if (!isSending) return;
 
     int r = red.toInt();
     int g = green.toInt();
@@ -57,96 +44,69 @@ class _SinglePixelControlPageState extends State<SinglePixelControlPage> {
     int p = brightness.toInt();
 
     String command;
-    // Mod kontrolünü boşlukları silip büyük harf olarak yapıyoruz.
-    bool isRGBW = (BLEHelper.mode.value.trim().toUpperCase() == "RGBW");
-    if (isRGBW) {
+    if (BLEHelper.mode.value == "RGBW") {
       int w = white.toInt();
-      command = "PIXCTRL:$pixelNumber:$r,$g,$b,$w,$p";
+      command = "RGBW:$r,$g,$b,$w,$p";
     } else {
-      command = "PIXCTRL:$pixelNumber:$r,$g,$b,$p";
+      command = "RGB:$r,$g,$b,$p";
     }
 
-    print("Gönderilen piksel komutu: $command");
-
+    List<int> bytes = command.codeUnits;
     if (BLEHelper.connectedCharacteristic != null) {
       try {
-        await BLEHelper.connectedCharacteristic!.write(command.codeUnits);
+        await BLEHelper.connectedCharacteristic!.write(bytes);
+        print("Gönderilen komut: $command");
       } catch (e) {
-        print("PIXCTRL komutu gönderme hatası: $e");
+        print("Renk komutu gönderme hatası: $e");
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text("PIXCTRL komutu gönderilemedi")),
+            const SnackBar(content: Text("Renk komutu gönderilemedi")),
           );
         }
+        // Hata alınırsa bağlı cihazı temizle.
+        BLEHelper.clearConnection();
       }
     } else {
+      print("Bağlı BLE cihazı yok!");
       if (mounted) {
         ScaffoldMessenger.of(
           context,
         ).showSnackBar(const SnackBar(content: Text("Bağlı BLE cihazı yok!")));
       }
+      BLEHelper.clearConnection();
     }
   }
 
   @override
-  void dispose() {
-    _debounce?.cancel();
-    _pixelController.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    // Mod kontrolü için büyük harfe çevirerek kontrol ediyoruz.
-    bool isRGBW = (BLEHelper.mode.value.trim().toUpperCase() == "RGBW");
     return Scaffold(
       backgroundColor: appColorsController.blackColor.value,
-      appBar: AppBar(
-        backgroundColor: appColorsController.blackColor.value,
-        // Geri tuşu eklenmiş durumda (AppBar'ın otomatik back buttonu çalışacaktır)
-        iconTheme: IconThemeData(color: appColorsController.whiteColor.value),
-        title: const Text(
-          "Tek Piksel Kontrol",
-          style: TextStyle(color: Colors.white),
-        ),
-      ),
       body: SafeArea(
         child: SingleChildScrollView(
           padding: const EdgeInsets.all(16.0),
           child: Column(
             children: [
-              const SizedBox(height: 20),
-              // Piksel numarası girilecek alan
-              TextField(
-                controller: _pixelController,
-                keyboardType: TextInputType.number,
-                style: TextStyle(color: appColorsController.whiteColor.value),
-                decoration: InputDecoration(
-                  labelText: "Piksel Numarası (1-indexli)",
-                  labelStyle: TextStyle(
-                    color: appColorsController.whiteColor.value,
-                  ),
-                  border: const OutlineInputBorder(),
-                ),
-              ),
-              const SizedBox(height: 20),
+              const SizedBox(height: 50),
               // Renk önizleme kutusu
-              Container(
-                width: double.infinity,
-                height: 150,
-                decoration: BoxDecoration(
-                  color: Color.fromARGB(
-                    brightness.toInt(),
-                    red.toInt(),
-                    green.toInt(),
-                    blue.toInt(),
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 24),
+                child: Container(
+                  width: double.infinity,
+                  height: 150,
+                  decoration: BoxDecoration(
+                    color: Color.fromARGB(
+                      brightness.toInt(),
+                      red.toInt(),
+                      green.toInt(),
+                      blue.toInt(),
+                    ),
+                    borderRadius: BorderRadius.circular(12),
+                    border: Border.all(color: Colors.white, width: 2),
                   ),
-                  borderRadius: BorderRadius.circular(12),
-                  border: Border.all(color: Colors.white, width: 2),
                 ),
               ),
-              const SizedBox(height: 30),
-              // Renk ve parlaklık slider'ları
+              const SizedBox(height: 40),
+              // Sliders bölümünü içeren Column
               Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
@@ -210,8 +170,8 @@ class _SinglePixelControlPageState extends State<SinglePixelControlPage> {
                       _onColorChanged();
                     },
                   ),
-                  // White slider, sadece RGBW modunda gösterilir
-                  if (isRGBW) ...[
+                  // White slider, sadece mod "RGBW" ise gösterilir
+                  if (BLEHelper.mode.value == "RGBW") ...[
                     Text(
                       "White: ${white.toInt()}",
                       style: const TextStyle(color: Colors.grey, fontSize: 18),
